refactor(script): split party/guild buildins into buildin_party.c
Issue #61 fourth step: move party, guild, agit/castle, and marriage/adoption helpers out of script.c.
